requirements.txt
pytelegrambotapi
pandas
requests
Pillow